reset mot de passe

+++ src/Controller/AdminController.php
@@ -7,6 +7,7 @@ use App\Entity\Ville;
 use App\Form\SiteType;
 use App\Form\VilleType;
 use App\Repository\SiteRepository;
+use App\Repository\UserRepository;
 use App\Repository\VilleRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -71,9 +72,17 @@ class AdminController extends AbstractController
     #[Route('/villes/suppr/{id}', name: 'villes_suppr')]
     public function supprVilles(VilleRepository $villeRepository, int $id, EntityManagerInterface $entityManager): Response
     {
+
         $ville = $villeRepository->find($id);
-        $entityManager->remove($ville);
-        $entityManager->flush();
+        $lieu = $ville->getLieu();
+        if ($lieu->isEmpty()) {
+            $entityManager->remove($ville);
+            $entityManager->flush();
+
+            $this->addFlash('success', 'Votre ville a été supprimée !');
+        } else {
+            $this->addFlash('error', 'Votre ville n\'a pas été supprimée car elle est rattachée à un lieu !');
+        }
 
         return $this->redirectToRoute('admin_villes', [
 
@@ -100,7 +109,6 @@ class AdminController extends AbstractController
 
         ]);
     }
-
 
 
     #[Route('/sites', name: 'sites')]
@@ -149,9 +157,19 @@ class AdminController extends AbstractController
     #[Route('/sites/suppr/{id}', name: 'sites_suppr')]
     public function supprSites(SiteRepository $siteRepository, int $id, EntityManagerInterface $entityManager): Response
     {
+
         $site = $siteRepository->find($id);
-        $entityManager->remove($site);
-        $entityManager->flush();
+
+        $utilisateurs = $site->getUsers();
+        $sorties = $site->getSorties();
+
+        if ($utilisateurs->isEmpty() and $sorties->isEmpty()) {
+            $entityManager->remove($site);
+            $entityManager->flush();
+        } else {
+            $this->addFlash('error', 'Votre site n\'a pas été supprimée car il est rattachée à un utilisateur ou à une sortie !');
+        }
+
 
         return $this->redirectToRoute('admin_sites', [
 
@@ -175,6 +193,70 @@ class AdminController extends AbstractController
 
         return $this->render('admin/creerSite.html.twig', [
             'siteForm' => $siteForm->createView(),
+
+        ]);
+    }
+
+    #[Route('/utilisateurs', name: 'utilisateurs')]
+    public function gererUtilisateurs(UserRepository $userRepository): Response
+    {
+        $users = $userRepository->findAll();
+
+        return $this->render('admin/utilisateurs.html.twig', [
+            'listeUser' => $users,
+        ]);
+    }
+
+    #[Route('/utilisateurs/tri', name: 'utilisateurs_tri')]
+    public function triUtilisateurs(UserRepository $userRepository, Request $request): Response
+    {
+        $barreRecherche = $request->get('rechercher');
+        $listeUsers = $userRepository->findByResearch($barreRecherche);
+
+
+        return $this->render('admin/utilisateurs.html.twig', [
+            'listeUser' => $listeUsers,
+        ]);
+    }
+
+    #[Route('/utilisateurs/desactive/{id}', name: 'utilisateur_desactive')]
+    public function desactiveUtilisateur(UserRepository $userRepository, int $id, EntityManagerInterface $entityManager, Request $request): Response
+    {
+        $user = $userRepository->find($id);
+        $user->setActif(false);
+        $user->setRoles(['']);
+        $entityManager->persist($user);
+        $entityManager->flush();
+
+        return $this->redirectToRoute('admin_utilisateurs', [
+
+        ]);
+    }
+
+    #[Route('/utilisateurs/active/{id}', name: 'utilisateur_active')]
+    public function activeUtilisateur(UserRepository $userRepository, int $id, EntityManagerInterface $entityManager, Request $request): Response
+    {
+        $user = $userRepository->find($id);
+        $user->setActif(true);
+        $user->setRoles(['ROLE_USER']);
+        $entityManager->persist($user);
+        $entityManager->flush();
+
+        return $this->redirectToRoute('admin_utilisateurs', [
+
+        ]);
+    }
+
+
+    #[Route('/utilisateurs/suppr/{id}', name: 'utilisateurs_suppr')]
+    public function supprUtilisateurs(UserRepository $userRepository, int $id, EntityManagerInterface $entityManager): Response
+    {
+        // qu'est ce qu'il se passe si on suppr l'utilisateur ? suppr des sorties qu'il a crée ?
+        $user = $userRepository->find($id);
+        $entityManager->remove($user);
+        $entityManager->flush();
+
+        return $this->redirectToRoute('admin_utilisateurs', [
 
         ]);
     }
